move proto to config and v2alpha
