+++ api/envoy/config/grpc_credentials/v2alpha/file_based_metadata.proto
@@ -3,8 +3,8 @@ syntax = "proto3";
 // [#protodoc-title: Grpc Credentials File Based Metadata]
 // Configuration for File Based Metadata Grpc Credentials Plugin
 
-package envoy.extensions.grpc_credentials;
-option go_package = "grpc_file_creds";
+package envoy.config.grpc_credentials.v2alpha;
+option go_package = "v2alpha";
 
 import "envoy/api/v2/core/base.proto";
