--- api/envoy/extensions/grpc_credentials/BUILD
@@ -1,9 +1,0 @@
-load("//bazel:api_build_system.bzl", "api_proto_library")
-
-licenses(["notice"])  # Apache 2
-
-api_proto_library(
-    name = "file_based_metadata",
-    srcs = ["file_based_metadata.proto"],
-    deps = ["//envoy/api/v2/core:base"],
-)
