+++ source/extensions/grpc_credentials/file_based_metadata/BUILD
@@ -21,6 +21,6 @@ envoy_cc_library(
         "//source/common/grpc:common_lib",
         "//source/common/grpc:google_grpc_creds_lib",
         "//source/extensions/grpc_credentials:well_known_names",
-        "@envoy_api//envoy/extensions/grpc_credentials:file_based_metadata_cc",
+        "@envoy_api//envoy/config/grpc_credentials/v2alpha:file_based_metadata_cc",
     ],
 )
