+++ source/extensions/grpc_credentials/file_based_metadata/config.cc
@@ -1,7 +1,7 @@
 #include "extensions/grpc_credentials/file_based_metadata/config.h"
 
 #include "envoy/api/v2/core/grpc_service.pb.h"
-#include "envoy/extensions/grpc_credentials/file_based_metadata.pb.validate.h"
+#include "envoy/config/grpc_credentials/v2alpha/file_based_metadata.pb.validate.h"
 #include "envoy/grpc/google_grpc_creds.h"
 #include "envoy/registry/registry.h"
 
@@ -36,7 +36,7 @@ FileBasedMetadataGrpcCredentialsFactory::getChannelCredentials(
             Envoy::Config::Utility::translateToFactoryConfig(
                 credential.from_plugin(), file_based_metadata_credentials_factory);
         const auto& file_based_metadata_config = Envoy::MessageUtil::downcastAndValidate<
-            const envoy::extensions::grpc_credentials::FileBasedMetadataConfig&>(
+            const envoy::config::grpc_credentials::v2alpha::FileBasedMetadataConfig&>(
             *file_based_metadata_config_message);
         std::shared_ptr<grpc::CallCredentials> new_call_creds = grpc::MetadataCredentialsFromPlugin(
             std::make_unique<FileBasedMetadataAuthenticator>(file_based_metadata_config));
