+++ test/extensions/grpc_credentials/file_based_metadata/BUILD
@@ -18,6 +18,6 @@ envoy_cc_test(
         "//source/extensions/grpc_credentials/file_based_metadata:config",
         "//test/common/grpc:grpc_client_integration_test_harness_lib",
         "//test/integration:integration_lib",
-        "@envoy_api//envoy/extensions/grpc_credentials:file_based_metadata_cc",
+        "@envoy_api//envoy/config/grpc_credentials/v2alpha:file_based_metadata_cc",
     ] + envoy_select_google_grpc(["//source/common/grpc:google_async_client_lib"]),
 )
